fix:(context): заменил контекст на Zustand

+++ Server/controllers/favoritesControllers.js
@@ -6,7 +6,7 @@ class FavorController {
             const userId = req.user.userId;
             const favorites = await favoriteService.getFavor(userId)
             if (!userId) {
-                return res.status(400).json({ error: 'User ID not found' });
+                return res.status(400).json({ error: 'Не передан ID пользователя' });
             }
             res.json(favorites || [])
         } catch (error) {
@@ -17,6 +17,9 @@ class FavorController {
     async addItem(req, res, next) {
         try {
             const userId = req.user.userId;
+            if (!userId) {
+                return res.status(400).json({ error: 'Не передан ID пользователя' });
+            }
             const { productId } = req.body
             const updatedFavorites = await favoriteService.addItem(userId, productId)
             res.json(updatedFavorites || [])
@@ -28,6 +31,9 @@ class FavorController {
     async removeItem(req, res, next) {
         try {
             const userId = req.user.userId;
+            if (!userId) {
+                return res.status(400).json({ error: 'Не передан ID пользователя' });
+            }
             const { productId } = req.params
             const updatedFavorites = await favoriteService.removeItem(userId, productId)
             res.json(updatedFavorites || [])
